Warning comparison (un)signed.
git-svn-id: http://root.cern.ch/svn/root/trunk@47526 27541ba8-7e3a-0410-8455-c3a389f83636

+++ lib/MetaProcessor/MetaProcessor.cpp
@@ -552,7 +552,7 @@ namespace cling {
           return Interpreter::kFailure;
         }
         unsigned printable = 0;
-        for (int i = 0; i < readMagic; ++i)
+        for (size_t i = 0; i < readMagic; ++i)
           if (isprint(magic[i]))
             ++printable;
         if (10 * printable <  5 * readMagic) {
